MAINT: refactored the assembler common methods to base class
